# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Discovery-маркетплейс специалистов: каталог, поиск, лиды, отзывы.",
        "title": "marketpclce API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/leads": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Менеджер/клиент создаёт лид и выбирает специалистов-получателей. В ответе — id и контакты выбранных спецов (видны только создателю).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "Создать заявку (lead)",
                "parameters": [
                    {
                        "description": "lead payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_leads.createReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.CreateResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/me/leads/incoming": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "Входящие заявки специалиста",
                "parameters": [
                    {
                        "type": "string",
                        "description": "sent|viewed|accepted|declined",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "default 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.incomingListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.errorResponse"
                        }
                    }
                }
            }
        },
        "/me/leads/{id}/recipient": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "leads"
                ],
                "summary": "Обновить статус получателя по лиду",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lead id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "статус",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_leads.recipientReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.recipientStatusResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_leads.errorResponse"
                        }
                    }
                }
            }
        },
...
        "/me/profile/check": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "LLM возвращает вердикты и подсказки по тексту bio и display_name. Доступен, только если включён LLM провайдер.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Проверить bio/имя профиля через LLM",
                "parameters": [
                    {
                        "description": "профильные поля",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_profilecheck.checkReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_profilecheck.Result"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_profilecheck.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_profilecheck.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/internal_profilecheck.errorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_profilecheck.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_leads.CreateResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "specialists": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_leads.SpecialistContact"
                    }
                }
            }
        },
        "internal_leads.IncomingLead": {
            "type": "object",
            "properties": {
                "brief": {
                    "type": "string"
                },
                "budget_max": {
                    "type": "integer"
                },
                "budget_min": {
                    "type": "integer"
                },
                "client_contact": {
                    "type": "string"
                },
                "client_name": {
                    "type": "string"
                },
                "client_user_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recipient_count": {
                    "type": "integer"
                },
                "recipient_responded_at": {
                    "type": "string"
                },
                "recipient_status": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_category_code": {
                    "type": "string"
                }
            }
        },
        "internal_leads.SpecialistContact": {
            "type": "object",
            "properties": {
                "contact_email": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "internal_leads.createReq": {
            "type": "object",
            "properties": {
                "brief": {
                    "type": "string"
                },
                "budget_max": {
                    "type": "integer"
                },
                "budget_min": {
                    "type": "integer"
                },
                "client_contact": {
                    "type": "string"
                },
                "client_name": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "specialist_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "target_category": {
                    "type": "string"
                }
            }
        },
        "internal_leads.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_leads.incomingListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_leads.IncomingLead"
                    }
                }
            }
        },
        "internal_leads.recipientReq": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_leads.recipientStatusResp": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "internal_profilecheck.PartResult": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                },
                "reasons": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "score": {
                    "type": "integer"
                },
                "suggestion": {
                    "type": "string"
                }
            }
        },
        "internal_profilecheck.Result": {
            "type": "object",
            "properties": {
                "bio": {
                    "$ref": "#/definitions/internal_profilecheck.PartResult"
                },
                "name": {
                    "$ref": "#/definitions/internal_profilecheck.PartResult"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "internal_profilecheck.checkReq": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                }
            }
        },
        "internal_profilecheck.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}